Move use-unmute-user file location
